# Conflicts:
#	codalab/common.py
#	worker/codalabworker/main.py
#	worker/codalabworker/run.py
#	worker/codalabworker/worker.py
#	worker/setup.py

+++ worker/codalabworker/formatting.py
@@ -53,16 +53,16 @@ def parse_size(s):
     """
     try:
         if s[-1].isdigit():
-            return float(s)
+            return int(s)
         n, unit = float(s[0:-1]), s[-1].lower()
         if unit == 'k':
-            return n * 1024
+            return int(n * 1024)
         if unit == 'm':
-            return n * 1024 * 1024
+            return int(n * 1024 * 1024)
         if unit == 'g':
-            return n * 1024 * 1024 * 1024
+            return int(n * 1024 * 1024 * 1024)
         if unit == 't':
-            return n * 1024 * 1024 * 1024 * 1024
+            return int(n * 1024 * 1024 * 1024 * 1024)
     except (IndexError, ValueError):
         pass  # continue to next line and throw error
     raise ValueError('Invalid size: %s, expected <number>[<k|m|g|t>]' % s)
